requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes -o requirements-dev.txt
-e .
appnope==0.1.4 ; sys_platform == 'darwin'
asttokens==3.0.0
certifi==2024.12.14
cffi==1.17.1 ; implementation_name == 'pypy'
charset-normalizer==3.4.1
colorama==0.4.6 ; sys_platform == 'win32'
comm==0.2.2
dateparser==1.2.0
debugpy==1.8.12
decorator==5.1.1
executing==2.2.0
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jupyter-client==8.6.3
jupyter-core==5.7.2
mako==1.3.8
markdown==3.7
markupsafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdoc3==0.11.5
pexpect==4.9.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
platformdirs==4.3.6
pluggy==1.5.0
prompt-toolkit==3.0.50
psutil==6.1.1
ptyprocess==0.7.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
pure-eval==0.2.3
pycparser==2.22 ; implementation_name == 'pypy'
pygments==2.19.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308 ; platform_python_implementation != 'PyPy' and sys_platform == 'win32'
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
wcwidth==0.2.13
